requirements.txt
Pillow~=7.0.0
pygame~=1.9.6
numpy~=1.18.1
PyGLM~=1.1.8
